Move TagsCloudVisualizerExtensions to separate file
