+++ TagsCloudContainer/Visualizer/TagsCloudVisualizer.cs
@@ -42,17 +42,4 @@ namespace TagsCloudContainer.Visualizer
             }
         }
     }
-
-    public static class TagsCloudVisualizerExtensions
-    {
-        public static byte[] ToByteArray(this Bitmap image)
-        {
-            using (var memoryStream = new MemoryStream())
-            {
-                image.Save(memoryStream, ImageFormat.Png);
-
-                return memoryStream.ToArray();
-            }
-        }
-    }
 }
